Replaced About/License activity with Fragments
