+++ app/src/main/AndroidManifest.xml
@@ -48,15 +48,8 @@
         <activity
             android:name=".About.AboutActivity"
             android:label="@string/about"
-            android:theme="@style/AppTheme.MaterialAboutActivity"
+            android:theme="@style/AppTheme"
             android:parentActivityName=".Main.TimerActivity">
-        </activity>
-
-        <activity
-            android:name=".About.LicencesActivity"
-            android:label="@string/about_open_source_licences"
-            android:theme="@style/AppTheme.MaterialAboutActivity"
-            android:parentActivityName=".About.AboutActivity">
         </activity>
 
         <activity
